+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 2.0.5
+* Make `@Expose()` in `Controller` optional. https://github.com/angel-dart/angel/issues/107
 * Add `allowHttp1` to `AngelHttp2` constructors. https://github.com/angel-dart/angel/issues/108
 * Add `deserializeBody` and `decodeBody` to `RequestContext`. https://github.com/angel-dart/angel/issues/109
 * Add `HostnameRouter`, which allows for routing based on hostname. https://github.com/angel-dart/angel/issues/110